Massive overhaul + additions
Completely overhauled most of the scan / target review code, which is
now working fairly well on my PC. Attack mode is buggy and is still a
work in progress (especially the output it generates), but I had no
problem setting up a rotation between 3 AP's with rate limiting. I
expect to have reavetard ready for beta testing and documentation in a
couple of days.